9719-move docket clerk adds unservable document to its own file
